requirements.txt
tensorflow==2.11.0
numpy==1.21.6
pandas==1.3.5
matplotlib==3.5.1
seaborn==0.11.2
scikit-learn==1.0.2
opencv-python==4.6.0.66
PyYAML==6.0
h5py==3.6.0
google-colab==1.0.0